backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embodiedai-mvp"
version = "0.1.0"
description = "Embodied AI MVP backend with ROS2, MuJoCo, and Rerun integration."
readme = "README.md"
authors = [{ name = "EmbodiedAI Team" }]
requires-python = ">=3.10"
dependencies = [
  "aiohttp>=3.9.5",
  "mujoco>=3.0.0",
  "numpy>=1.26",
  "opencv-python>=4.10",
  "pydantic>=2.8",
  "PyYAML>=6.0",
  "rerun-sdk[web]>=0.15.0",
  "rosbags>=0.10.3",
  "transforms3d>=0.4.2",
  "typer>=0.12.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "ruff>=0.6.9"
]
ros = []

[project.scripts]
embodiedai-sim = "embodiedai_mvp.mvp_nodes.sim_node:main"
embodiedai-vision = "embodiedai_mvp.mvp_nodes.vision_node:main"
embodiedai-grasp = "embodiedai_mvp.mvp_nodes.grasp_action_server:main"
embodiedai-mcp = "embodiedai_mvp.mcp_server.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["embodiedai_mvp*"]

[tool.ruff]
line-length = 100
extend-select = ["B", "C4", "I", "N", "Q", "RET", "RUF100", "TID"]

[tool.pytest.ini_options]
addopts = "-ra"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
